Add files via upload
Create an interactive navigation
menu that changes color or style
when scrolled or when hovering
over a menu item. The navigation
menu should have a fixed position
and be visible on all pages. Use
HTML to structure the menu, CSS to
style it, and JavaScript to add
interactivity, such as changing the
background color or font color of
the menu when it is scrolled or
when a menu item is hovered over.